Interpret Valid Sequence or  Decode Ways.
HackerRank Interview Question. 

Consider a code system where alphabets are represented as sequential decimal numbers starting from 0. 
E.g.
'A' -> 0
'B' -> 1
...
'Z' -> 25

Given a string of digits e.g. "123" as input, print out the number of possible valid representations of letters.
For example, "123" can be interpreted as "bcd " (as "1","2","3"), or "bx" (as "1","23") or "md" (as "12", "3").

So we have 3 valid interpretations of the string "123".